media-source/SourceBuffer-remove_invalid-end.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#remove(start, invalidEnd)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm': 'video/webm; codecs="vorbis,vp8"',
                '/media/segmented-3s.mp4' : 'video/mp4'};

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function testMedia(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updateend', function(e) {
                    t.step(function() {
                        assert_throws({name: 'InvalidAccessError'},
                                      function() {
                                          sourceBuffer.remove(1, 0);
                                      },
                                      'SourceBuffer#remove(start, less_than_start)');
                        assert_throws({name: 'InvalidAccessError'},
                                      function() {
                                          sourceBuffer.remove(1, 1);
                                      },
                                      'SourceBuffer#remove(start, equal_to_start)');
                    });
                    t.done();
                });
                sourceBuffer.appendBuffer(data);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'SourceBuffer#remove(start, invalidEnd) (' + mime + ') : ' +
       'If end is invalid, InvalidAccessError should be thrown');
}
for(var file in contents) {
    testMedia(file, contents[file]);
}
</script>
</body>
</html>
